add GroupHelper
