+++ test/test_add_group.py
@@ -13,11 +13,11 @@ def app(request):
 
 def test_add_group(app):
     app.session.login(username="admin", password="secret")
-    app.create_group(Group(name="namegroupqa1", header="text", footer="texte"))
+    app.group.create(Group(name="namegroupqa1", header="text", footer="texte"))
     app.session.logout()
 
 
 def test_add_empty_group(app):
     app.session.login(username="admin", password="secret")
-    app.create_group(Group(name="", header="", footer=""))
+    app.group.create(Group(name="", header="", footer=""))
     app.session.logout()
